Cleanup duplicated functions for a more DRY codebase
