+++ wagtail/admin/views/account.py
@@ -5,10 +5,8 @@ from django.contrib import messages
 from django.contrib.auth import views as auth_views
 from django.contrib.auth import update_session_auth_hash
 from django.contrib.auth.forms import PasswordChangeForm
-from django.core.exceptions import ImproperlyConfigured
 from django.http import Http404
 from django.shortcuts import redirect, render
-from django.utils.module_loading import import_string
 from django.utils.translation import ugettext as _
 from django.utils.translation import activate
 from django.views.decorators.cache import never_cache
@@ -19,17 +17,7 @@ from wagtail.admin.utils import get_available_admin_languages
 from wagtail.core.models import UserPagePermissionsProxy
 from wagtail.users.forms import NotificationPreferencesForm, PreferredLanguageForm
 from wagtail.users.models import UserProfile
-
-
-# Helper functions to allow customisation of the user login form
-def get_custom_form(form_setting):
-    try:
-        return import_string(getattr(settings, form_setting))
-    except ImportError:
-        raise ImproperlyConfigured(
-            "%s refers to a form '%s' that is not available" %
-            (form_setting, getattr(settings, form_setting))
-        )
+from wagtail.utils.loading import get_custom_form
 
 
 def get_user_login_form():
@@ -56,7 +44,8 @@ def password_reset_enabled():
 
 def account(request):
     user_perms = UserPagePermissionsProxy(request.user)
-    show_notification_preferences = user_perms.can_edit_pages() or user_perms.can_publish_pages()
+    show_notification_preferences = user_perms.can_edit_pages(
+    ) or user_perms.can_publish_pages()
 
     return render(request, 'wagtailadmin/account/account.html', {
         'show_change_password': password_management_enabled() and request.user.has_usable_password(),
@@ -79,7 +68,8 @@ def change_password(request):
                 form.save()
                 update_session_auth_hash(request, form.user)
 
-                messages.success(request, _("Your password has been changed successfully!"))
+                messages.success(request, _(
+                    "Your password has been changed successfully!"))
                 return redirect('wagtailadmin_account')
         else:
             form = PasswordChangeForm(request.user)
@@ -103,20 +93,24 @@ def _wrap_password_reset_view(view_func):
 
 password_reset = _wrap_password_reset_view(auth_views.password_reset)
 password_reset_done = _wrap_password_reset_view(auth_views.password_reset_done)
-password_reset_confirm = _wrap_password_reset_view(auth_views.password_reset_confirm)
-password_reset_complete = _wrap_password_reset_view(auth_views.password_reset_complete)
+password_reset_confirm = _wrap_password_reset_view(
+    auth_views.password_reset_confirm)
+password_reset_complete = _wrap_password_reset_view(
+    auth_views.password_reset_complete)
 
 
 def notification_preferences(request):
     if request.method == 'POST':
-        form = NotificationPreferencesForm(request.POST, instance=UserProfile.get_for_user(request.user))
+        form = NotificationPreferencesForm(
+            request.POST, instance=UserProfile.get_for_user(request.user))
 
         if form.is_valid():
             form.save()
             messages.success(request, _("Your preferences have been updated."))
             return redirect('wagtailadmin_account')
     else:
-        form = NotificationPreferencesForm(instance=UserProfile.get_for_user(request.user))
+        form = NotificationPreferencesForm(
+            instance=UserProfile.get_for_user(request.user))
 
     # quick-and-dirty catch-all in case the form has been rendered with no
     # fields, as the user has no customisable permissions
@@ -130,7 +124,8 @@ def notification_preferences(request):
 
 def language_preferences(request):
     if request.method == 'POST':
-        form = PreferredLanguageForm(request.POST, instance=UserProfile.get_for_user(request.user))
+        form = PreferredLanguageForm(
+            request.POST, instance=UserProfile.get_for_user(request.user))
 
         if form.is_valid():
             user_profile = form.save()
@@ -140,7 +135,8 @@ def language_preferences(request):
             messages.success(request, _("Your preferences have been updated."))
             return redirect('wagtailadmin_account')
     else:
-        form = PreferredLanguageForm(instance=UserProfile.get_for_user(request.user))
+        form = PreferredLanguageForm(
+            instance=UserProfile.get_for_user(request.user))
 
     return render(request, 'wagtailadmin/account/language_preferences.html', {
         'form': form,
