+++ wagtail/users/views/users.py
@@ -8,6 +8,7 @@ from django.utils.module_loading import import_string
 from django.utils.translation import ugettext as _
 from django.views.decorators.vary import vary_on_headers
 
+from wagtail.utils.loading import get_custom_form
 from wagtail.utils.pagination import paginate
 from wagtail.admin import messages
 from wagtail.admin.forms import SearchForm
@@ -28,20 +29,10 @@ change_user_perm = "{0}.change_{1}".format(AUTH_USER_APP_LABEL, AUTH_USER_MODEL_
 delete_user_perm = "{0}.delete_{1}".format(AUTH_USER_APP_LABEL, AUTH_USER_MODEL_NAME.lower())
 
 
-def get_custom_user_form(form_setting):
-    try:
-        return import_string(getattr(settings, form_setting))
-    except ImportError:
-        raise ImproperlyConfigured(
-            "%s refers to a form '%s' that is not available" %
-            (form_setting, getattr(settings, form_setting))
-        )
-
-
 def get_user_creation_form():
     form_setting = 'WAGTAIL_USER_CREATION_FORM'
     if hasattr(settings, form_setting):
-        return get_custom_user_form(form_setting)
+        return get_custom_form(form_setting)
     else:
         return UserCreationForm
 
@@ -49,7 +40,7 @@ def get_user_creation_form():
 def get_user_edit_form():
     form_setting = 'WAGTAIL_USER_EDIT_FORM'
     if hasattr(settings, form_setting):
-        return get_custom_user_form(form_setting)
+        return get_custom_form(form_setting)
     else:
         return UserEditForm
